added laravel breeze auth

+++ app/Http/Controllers/NoteController.php
@@ -12,7 +12,7 @@ class NoteController extends Controller
      */
     public function index()
     {
-        $notes = Note::query()->orderBy('created_at', 'desc')->paginate(10);
+        $notes = Note::query()->where('user_id', request()->user()->id)->orderBy('created_at', 'desc')->paginate(10);
         return view('note.index', ['notes' => $notes]);
     }
 
@@ -33,7 +33,7 @@ class NoteController extends Controller
             'note' =>  ['required', 'string']
         ]);
 
-        $data['user_id'] = 1;
+        $data['user_id'] = $request->user()->id;
         $note = Note::create($data);
 
         return to_route('note.show', $note)->with('message', 'Note was create');
@@ -44,6 +44,10 @@ class NoteController extends Controller
      */
     public function show(Note $note)
     {
+        if ($note->user_id !== request()->user()->id) {
+            abort(403);
+        }
+
         return view('note.show', ['note' => $note]);
     }
 
@@ -52,7 +56,10 @@ class NoteController extends Controller
      */
     public function edit(Request $request, Note $note)
     {
-        return view('note.edit');
+        if ($note->user_id !== request()->user()->id) {
+            return abort(403);
+        }
+        return view('note.edit', ['note' => $note]);
     }
 
     /**
@@ -60,6 +67,9 @@ class NoteController extends Controller
      */
     public function update(Request $request, Note $note)
     {
+        if ($note->user_id !== request()->user()->id) {
+            return abort(403);
+        }
         $data = $request->validate([
             'note' =>  ['required', 'string']
         ]);
@@ -75,6 +85,9 @@ class NoteController extends Controller
      */
     public function destroy(Note $note)
     {
+        if ($note->user_id !== request()->user()->id) {
+            return abort(403);
+        }
         $note->delete();
 
         return to_route('note.index', $note)->with('message', 'Note was deleted');
